修复课时切换与课程附件导出并优化判题派发

+++ lms/lms/judge_service.py
@@ -102,7 +102,8 @@ def submit_programming_exercise(
 	frappe.enqueue(
 		"lms.lms.judge_service.dispatch_submission",
 		submission_name=doc.name,
-		queue="long",
+		# Dispatch is a short HTTP request; never wait behind course indexing/imports.
+		queue="short",
 		enqueue_after_commit=True,
 	)
 	return {"submission": doc.name, "status": doc.status}
